requirements.txt
// use o comando pip install requirements.txt para baixar todas as bibliotecas do projeto
pymodbus>=3.3.2
APScheduler>=3.11.2
streamlit>=1.26.0
pandas>=1.5.3
fpdf2>=2.7.0
pytest>=7.0.0
python-dotenv
